move: region을 post 하위로 이동
