+++ lib/features/post/domain/use_case/post_service.dart
@@ -1,4 +1,4 @@
-import 'package:sodong_app/features/post_list/domain/models/region.dart';
+import 'package:sodong_app/features/post/domain/entities/region.dart';
 import 'package:sodong_app/features/post_list/domain/models/town_life_post.dart';
 import 'package:sodong_app/features/post_list/domain/repositories/post_repository.dart';
