+++ lib/features/post/presentation/pages/post_list/widgets/region_selector.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:sodong_app/features/location/location_viewmodel.dart';
-import 'package:sodong_app/features/post_list/domain/models/region.dart';
+import 'package:sodong_app/features/post/domain/entities/region.dart';
 import 'package:sodong_app/features/post_list/presentation/view_models/region_view_model.dart';
 
 class RegionSelector extends ConsumerWidget {
