+++ lib/features/post_list/presentation/view_models/pagination_manager.dart
@@ -1,5 +1,5 @@
+import 'package:sodong_app/features/post/domain/entities/region.dart';
 import 'package:sodong_app/features/post/domain/use_case/post_service.dart';
-import 'package:sodong_app/features/post_list/domain/models/region.dart';
 import 'package:sodong_app/features/post_list/domain/models/town_life_post.dart';
 
 /// 페이지네이션 관리 클래스
